After a small change of the metor.ieee.org pages, DociiMentor was not able to scrape the info off of those pages. This has now been fixed.
Realizing that, depending on configuration of the computer, mapping of
directories, etc., it may not be a good idea to place the .dociimentor
directory, which holds the configuration file, under the home directory.
It is now placed in the current working directory, i.e., the directory
from which DociiMentor is launched.